refactor(utils): add shared toFiniteNumber for metadata discount coercion
