+++ packages/utils/src/functions/index.ts
@@ -1,6 +1,7 @@
 export * from './array';
 export * from './currency';
 export * from './fileValidation';
+export * from './number';
 export * from './sanitize';
 export * from './slug';
 export * from './transcript-vtt';
