+++ packages/utils/src/functions/number.ts
@@ -1,0 +1,18 @@
+/**
+ * Coerces unknown input to a finite number.
+ * String numbers are parsed; empty, null, and invalid values return undefined.
+ */
+export function toFiniteNumber(value: unknown): number | undefined {
+  if (typeof value === 'number' && Number.isFinite(value)) {
+    return value;
+  }
+
+  if (typeof value === 'string' && value.trim()) {
+    const parsed = Number(value);
+    if (Number.isFinite(parsed)) {
+      return parsed;
+    }
+  }
+
+  return undefined;
+}
